@@ -1,23 +1,25 @@
 plugins {
+    id 'kotlin-kapt'
     id 'com.android.application'
     id 'kotlin-android'
-    id 'kotlin-android-extensions'
-    id 'kotlin-kapt'
     id 'dagger.hilt.android.plugin'
 }
 
 android {
-    compileSdkVersion 29
-    buildToolsVersion "30.0.1"
+    namespace = "pl.paullettuce.android_astarium_interview_app"
+    compileSdk = 34
 
     defaultConfig {
-        applicationId "pl.paullettuce.android_astarium_interview_app"
-        minSdkVersion 21
-        targetSdkVersion 29
-        versionCode 1
-        versionName "1.0"
+        applicationId = "pl.paullettuce.android_astarium_interview_app"
+        minSdk = 21
+        targetSdk = 34
+        versionCode = 1
+        versionName = "1.0"
 
-        testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
+        testInstrumentationRunner = "androidx.test.runner.AndroidJUnitRunner"
+        vectorDrawables {
+            useSupportLibrary = true
+        }
     }
 
     buildTypes {
@@ -27,24 +29,28 @@ android {
         }
     }
     compileOptions {
-        sourceCompatibility JavaVersion.VERSION_1_8
-        targetCompatibility JavaVersion.VERSION_1_8
+        sourceCompatibility = JavaVersion.VERSION_17
+        targetCompatibility = JavaVersion.VERSION_17
     }
     kotlinOptions {
-        jvmTarget = '1.8'
+        jvmTarget = "17"
+    }
+
+    buildFeatures {
+        dataBinding = true
     }
 }
 
 dependencies {
     //Rx
-    def rx_version = "3.0.9"
+    def rx_version = "3.1.5"
     implementation "io.reactivex.rxjava3:rxjava:$rx_version"
     implementation "io.reactivex.rxjava3:rxkotlin:3.0.1"
-    implementation 'io.reactivex.rxjava3:rxandroid:3.0.0'
+    implementation 'io.reactivex.rxjava3:rxandroid:3.0.2'
 
     // Dagger - Hilt
-    implementation "com.google.dagger:hilt-android:2.28-alpha"
-    kapt "com.google.dagger:hilt-android-compiler:2.28-alpha"
+    implementation "com.google.dagger:hilt-android:2.48"
+    kapt "com.google.dagger:hilt-android-compiler:2.48"
 
     // Retrofit
     def retrofit_version = "2.9.0"
