@@ -1,11 +1,9 @@
 package pl.paullettuce.android_astarium_interview_app.presentation.stations_distance.list
 
-import android.view.View
+import android.view.LayoutInflater
 import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
-import kotlinx.android.synthetic.main.list_item_station_info.view.*
-import pl.paullettuce.android_astarium_interview_app.R
-import pl.paullettuce.android_astarium_interview_app.domain.extensions.inflate
+import pl.paullettuce.android_astarium_interview_app.databinding.ListItemStationInfoBinding
 import pl.paullettuce.android_astarium_interview_app.domain.model.StationInfo
 import pl.paullettuce.android_astarium_interview_app.presentation.stations_distance.StationsDistanceContract
 
@@ -15,13 +13,15 @@ class StationInfoListAdapter(
     private val items = mutableListOf<StationInfo>()
 
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): StationInfoViewHolder {
-        return StationInfoViewHolder(parent.inflate(R.layout.list_item_station_info), interactor)
+        val binding =
+            ListItemStationInfoBinding.inflate(LayoutInflater.from(parent.context), parent, false)
+        return StationInfoViewHolder(binding, interactor)
     }
 
     override fun getItemCount() = items.size
 
     override fun onBindViewHolder(holder: StationInfoViewHolder, position: Int) {
-         holder.bind(items[position])
+        holder.bind(items[position])
     }
 
     fun setItems(newItems: List<StationInfo>) {
@@ -34,13 +34,13 @@ class StationInfoListAdapter(
 }
 
 class StationInfoViewHolder(
-    itemView: View,
+    private val binding: ListItemStationInfoBinding,
     private val interactor: StationsDistanceContract.StationInfoListInteractor
-) : RecyclerView.ViewHolder(itemView) {
+) : RecyclerView.ViewHolder(binding.root) {
     fun bind(item: StationInfo) {
-        itemView.title.text = item.name
-        itemView.subtitle.text = item.regionInfo
-        itemView.initials.text = item.initials
-        itemView.setOnClickListener { interactor.onStationInfoListItemClick(item) }
+        binding.title.text = item.name
+        binding.subtitle.text = item.regionInfo
+        binding.initials.text = item.initials
+        binding.root.setOnClickListener { interactor.onStationInfoListItemClick(item) }
     }
 }
